Extract server-only data fetching to route.server.ts
